refactor(docs): extract SDK reference into lib/sdk-reference.ts
Pure relocation of the @makinbakin/sdk reference generator out of
scripts/docs/generate.ts: the TypeScript-compiler-API symbol/JSDoc extraction
(readSdkExports, extractSdkSymbols, jsdocTextOf, hasExportModifier,
inferKindFromName, extractInterfaceMembers), the SdkSymbol/SdkMember/SdkSubpath
types, the CORE_TYPES/HOOKS_GROUPS/TYPE_DOMAIN_GROUPS grouping consts, the
sdk.md renderers (renderSdkReference + the six per-subpath helpers), and
validateSdkCoverage. Function bodies byte-for-byte unchanged; generate.ts
imports renderSdkReference + readSdkExports (the latter still consumed by the
coverage report + the sdk-reference llm bundle). Dropped the now-unused
typescript/relativeSource/escapeMd imports. docs:generate output verified
byte-identical.
